Testing the user controller

+++ backend/src/modules/user/user.services.ts
@@ -1,19 +1,13 @@
 import { Types } from "mongoose";
 import { ApiError } from "../../utils/apiError.js";
 import { User } from "../auth/auth.model.js";
-
-type registerServiceProps = {
-  name: string;
-  email: string;
-  password: string;
-}
+import { Transaction } from "../transaction/transaction.model.js";
 
 export const getAllUserService = async() => {
   try {
     const users = await User.find();
     return users;
   } catch(error: any) {
-    console.error("Transaction aborted:\n", error);
     throw new ApiError(error.statusCode || 500, error.message || "Failed in creating user");
   }
 } 
@@ -23,7 +17,6 @@ export const getUserService = async(id: string) => {
     const user = await User.findById(id);
     return user;
   } catch(error: any) {
-    console.error("Transaction aborted:\n", error);
     throw new ApiError(error.statusCode || 500, error.message || "Failed in creating user");
   }
 } 
@@ -40,18 +33,31 @@ export const updateUserRoleService = async(id: string, role: "USER" | "ADMIN") =
 
     return user;
   } catch(error: any) {
-    console.error("Transaction aborted:\n", error);
     throw new ApiError(error.statusCode || 500, error.message || "Failed in creating user");
   }
 } 
 
 export const deleteUserService = async(id: string) => {
   try {
-    const deletedUser = await User.findByIdAndDelete(id);
+    const transaction = await Transaction.findOne({
+      createdBy: new Types.ObjectId(id)
+    });
+
+    let deletedUser;
+    if (transaction) {
+      deletedUser = await User.findById(id);
+      if (deletedUser && deletedUser.isActive) {
+        deletedUser.isActive = false;
+        await deletedUser.save();
+      } else {
+        throw new ApiError(404, "User not found");
+      }
+    } else {
+      deletedUser = await User.findByIdAndDelete(id);
+    }
 
     return deletedUser;
   } catch(error: any) {
-    console.error("Transaction aborted:\n", error);
     throw new ApiError(error.statusCode || 500, error.message || "Failed in creating user");
   }
 }
